Make hacky symlink to `librealsense2` headers
CMake strips out the `/usr/include` path as compilers should add that as needed
anyways. However the Anaconda compilers do not do this as they are trying not
to use things from the system. Manually adding `/usr/include` to the headers
via `CXXFLAGS` causes really weird compiler errors like issues with `timespec`
from `time.h`. So that's not really viable. Hacking in this symlink seems to
get us past that issue and allows the build to proceed. Though it is really
hacky.